requirements.txt
numpy==2.2.4
pandas==2.2.3
Pillow==11.1.0
plotly==6.0.1
